+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Added default value parameter to RouteMatchInterface::getParameter() method to return when no value found for name.
 - Added Executor::execute() parameter lookup inside request attributes.
 - Added ContainerInterface::find() to find a value for path, returns default when not found.
+- Added default option for MultipleChoicePrompt::prompt() to return when no input is given.
 
 ### Changed